+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 * Rework .net build to be mostly automatic using nuget
 * Change .net target to more modern .net 4.0 framework
+* [#309](https://github.com/dblock/waffle/pull/309): Added impersonation support on spring-security filters [@sergey-podolsky](https://github.com/sergey-podolsky).
 * [#296](https://github.com/dblock/waffle/pull/296): Added Tomcat 9 support.
 * [#268](https://github.com/dblock/waffle/pull/301): Cannot log in automatically on machine where Tomcat service is running
 * [#274](https://github.com/dblock/waffle/pull/274): Update WindowsSecurityContextImpl.java to handle SEC_E_BUFFER_TOO_SMALL